requirements.txt
appdirs==1.4.4
colorama==0.4.4
distlib==0.3.1
filelock==3.0.12
future==0.18.2
mpmath==1.2.1
Pillow==8.1.0
pyglet==1.5.0
python-util==1.2.1
python-utils==2.5.6
scipy==1.6.1
six==1.15.0
sympy==1.9
termcolor==1.1.0
utils==1.0.1
virtualenv==20.4.2
